AI_Clinical_Trial_Matching_System/requirements.txt
# requirements.txt
pandas
requests
aiohttp
python-dotenv
sentence-transformers
scikit-learn
numpy
sqlalchemy
